Add --options argument to pass in all of the options as a JSON obj
In addition to the command line options, we now allow passing in the
options as a JSON obj using the --options argument. Arguments specified
in this way will be overridden by the same argument given as a normal
argument. For example, if given `-o '{"device-type":"trezor"}' -d coldcard`
args.device_type would be 'coldcard'.